requirements.txt
dnspython==2.2.1
cloudscraper==1.2.58
requests==2.27.1
